car_search/requirements.txt
Django==3.2.4
djangorestframework==3.12.4
psycopg2==2.9.1
geopy==2.2.0
djangorestframework==3.12.4
psycopg2-binary==2.9.1
geopy==2.2.0